requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
marshmallow
marshmallow-sqlalchemy
Flasgge
python-dotenv
